requirements.txt
Pyrogram
TgCrypto
youtube_dl
gTTS
pyaml
youtube-search-python
python-mpv